Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > CONFIG_AIO is on by default. I'd say it should be required -- does the lack of turning it off in the base configs communicate that?
> 
> No the default state of configs in the kernel doesn't do anything as far as establishing a requirement. If we want to mandate that all Android devices have the config enabled it needs to go into android-base.cfg.

I remember having issues with this before. I think the issue was that because CONFIG_AIO is on by default, it is not actually valid to have it in a defconfig since it generates conflicts between the final config and the defconfig. Since defconfigs add on to the base config, there might be similar problems with adding it to the base config.

Patch-set: 1
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0