Update patch set 1
Patch Set 1:

> I remember having issues with this before. I think the issue was that because CONFIG_AIO is on by default, it is not actually valid to have it in a defconfig since it generates conflicts between the final config and the defconfig. Since defconfigs add on to the base config, there might be similar problems with adding it to the base config.

This doesn't sound right to me... Device defconfigs are generated from the android-base.cfg:

https://source.android.com/devices/architecture/kernel/config

Once that is done, if there are config options in the generated defconfig that are default values for the architecture, those would disappear when a minimal defconfig is generated from the full defconfig, since they are redundant. But it is still necessary to have them in the android-base.cfg (assuming we are requiring them) since that is used for compliance testing in VTS and during boot time compatibility checking.

Patch-set: 1
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>