Update patch set 1
Patch Set 1:

Also, I am worried that because of b/111125947, when a device launch with P receives a full-OTA to Q (both Q system and Q vendor), OTA may report an error because the running kernel configuration (which sets xt_qtaguid) is checked against the new requirements (which requires xt_qtaguid to be unset). Maybe we need to put new kernel configs to boot.xml for OTA as well?

Patch-set: 1