add missing configs
CONFIG_AEABI is required on ARM as a prerequisite for
CONFIG_SECCOMP_FILTER.

CONFIG_ARM_KERNMEM_PERMS is ARM only and as such is moved to a new
ARM-only recommended fragment.

CONFIG_ARM64_SW_TTBR0_PAN is already in the ARM64 base fragment for 4.4
and 4.9 and can be removed from the recommended fragment in those
versions. On 4.14 it needs to be moved to an ARM64-specific recommended
fragment.

CONFIG_BINFMT_ELF is required as Android makes use of ELF binaries.

CONFIG_COMPAT is required on ARM64 platforms to enable
emulation/deprecated features.

CONFIG_COREDUMP and CONFIG_ELF_CORE are recommended for diagnostic
purposes.

CONFIG_DEVKMEM is no longer available for ARM64 on 4.14 and must be
moved into architecture-specific fragments.

CONFIG_MMU is required by ANDROID_BINDER_IPC.

CONFIG_RANDOMIZE_BASE and CONFIG_RELOCATABLE are not supported on ARM,
so are moved to arm64 and x86 recommended fragments.

CONFIG_TTY is a dependency for pty support which is required by
Android.

CONFIG_UPROBE_EVENT is only currently supported on x86 and ARM, not
ARM64, so it is moved to an architecture-specific recommendation.

Change-Id: I309c0cf969283f8781f70c39ccdb0fe261f23a31
Signed-off-by: Steve Muckle <smuckle@google.com>