Update patch set 3
Patch Set 3: -Code-Review

Sam, do you think this should really be in 'base' and not 'recommended'? We're not going to remove the old lac/pns patches, and in fact for P compatibility we need to cherry-pick those ppp drivers to 4.14 too.

But I solidly agree that we should deprecate this method for lts-next and I think it's a good idea to suggest to enabling the l2tp/pptp alternative in older kernels for testing on AOSP master.

Patch-set: 3
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0