add script to convert conditional requirements in xml
The xml file used to express conditional kernel config requirements
differs from the format used by the compatibility matrix in two ways.
This script fixes up the xml file so it may be included in the
compatibility matrix.

1. The kernel config XML specifies the minimum LTS version once at the
top of the file rather than for every group of configs in the file. It
also uses the "group" tag for groups of configs rather than the "kernel"
tag. The script removes the top entry and converts each "group" tag to a
"kernel" tag with the minimum LTS version.

2. The kernel config XML supports identifying config values as bools as
well as tristates. These values are treated no differently by the vintf
code, however it may reduce confusion for developers maintaining the
kernel configs by not forcing them to denote boolean kernel config
values as tristates in the XML source. These get converted to tristate
when creating the compatibility matrix as only tristate is supported
there.

Change-Id: Ifa4873c0a7d6c93375050a73fb56e9e3756c2c1b