Update patch set 3
Patch Set 3:

I think it would be better if the script had long-options for every conditional option and the user had to explicitly specify the choice for each. Relying on the values set in the platform defconfig creates a chicken-and-egg problem for bring up a new platform. The user also might have made a suboptimal choice in their platform defconfig and did not even realize it.

Patch-set: 3
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>