{o,o-mr1,p}/android-4.9+: relax QTAGUID requirements
Q userspace will use eBPF for bandwidth measurement
and does not require QTAGUID netfilter extension.
As such we are okay with either QTAGUID or the less
functional but sufficient OWNER extension (provided
the requisite eBPF config options are also enabled).

ie. we require either XT_MATCH_QTAGUID or XT_MATCH_OWNER.

Furthermore XT_MATCH_OWNER also requires:
  CONFIG_BPF_SYSCALL=y
  CONFIG_CGROUP_BPF=y
  CONFIG_NETFILTER_XT_MATCH_BPF=y
  CONFIG_NET_CLS_BPF=y
(the first 3 of which are already base requirements on P 4.9+)

Note: while this change is complex, it *widens* the
set of valid kernel configurations.  Any kernel that passed
configuration requirements before this change will also pass
them after this change.  However additionally this allows a
Q kernel to pass on a shipped as O or P but then upgraded to Q
device.

Bug: 125362907
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Ia6ce9e857d260cc6c74dd59731555479f7d38078