Update patch set 3
Patch Set 3:

The kernel configs in this repo are not directly used during the kernel build process for any particular device (at least, not that I am aware of). They are used directly by a VTS test which enforces that these kernel config requirements are met, and likewise, by a runtime check during Android boot that also enforces the requirements are met (more information in the README.md at the repo's top level).

I'm hesitant to automate the process of converting these requirement kernel config fragments into a platform defconfig. As Alistair pointed out there are requirements in the conditional config XML where the platform owner has a choice to make. Platform owners should be quite familiar with how their kernel is configured and the consequences of that configuration.

There are instructions in the README.md that show how to use the merge_config.sh script in the kernel tree to combine the various *.config fragments with the platform defconfig. Those instructions also prompt the platform owner to review the android-base-conditional.xml file if there is one and manually edit the defconfig to satisfy applicable requirements. The XML files are not too big, so this doesn't seem to me to be too much work for what is really a critical system configuration step.

Patch-set: 3
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>