Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> The kernel configs in this repo are not directly used during the kernel build process for any particular device (at least, not that I am aware of). They are used directly by a VTS test which enforces that these kernel config requirements are met, and likewise, by a runtime check during Android boot that also enforces the requirements are met (more information in the README.md at the repo's top level).
> 
> I'm hesitant to automate the process of converting these requirement kernel config fragments into a platform defconfig. As Alistair pointed out there are requirements in the conditional config XML where the platform owner has a choice to make. Platform owners should be quite familiar with how their kernel is configured and the consequences of that configuration.
> 
> There are instructions in the README.md that show how to use the merge_config.sh script in the kernel tree to combine the various *.config fragments with the platform defconfig. Those instructions also prompt the platform owner to review the android-base-conditional.xml file if there is one and manually edit the defconfig to satisfy applicable requirements. The XML files are not too big, so this doesn't seem to me to be too much work for what is really a critical system configuration step.


Thanks for kind explanation!
So this sounds like the kernel configs repository is mainly used to check that the platform defconfig file(like arch/arm64/configs/hikey_defconfig) in the vts test and boot process. If there is any failures reported, then we should update that hikey_defconfig file. Instead we should not use it to generate .config for building the kernel, which would not report the vts failure(if there was one with using only hikey_defconfig) as the kernel/configs repository is used already?

In our linaro lkft builds, we use the merge_config.sh(as in the README.md) to generate the .config with fragment configs in kernel/configs for the kernel building. I am not sure if we are doing the right way now.:(

Patch-set: 4
Reviewer: Gerrit User 1041813 <1041813@85c56323-6fa9-3386-8a01-6480fb634889>