Update patch set 4
Patch Set 4:

> So this sounds like the kernel configs repository is mainly used to check that the platform defconfig file(like arch/arm64/configs/hikey_defconfig) in the vts test and boot process. If there is any failures reported, then we should update that hikey_defconfig file. Instead we should not use it to generate .config for building the kernel, which would not report the vts failure(if there was one with using only hikey_defconfig) as the kernel/configs repository is used already?
> 
> In our linaro lkft builds, we use the merge_config.sh(as in the README.md) to generate the .config with fragment configs in kernel/configs for the kernel building. I am not sure if we are doing the right way now.:(

Having contributed to that README file, I always expected platform owners to merge the config fragments/platform defconfig once and then review them afterwards. It never really occurred to me that the process would be automated as part of the build.

Blindly merging in android-base.config is fine since there's no choice there - those options are required for Android. But the recommended configs really should be reviewed prior to merging. The conditional configs have to be reviewed for the correct selections to be made for the platform in question.

My advice would be to create and maintain a config fragment for your platform that contains the Android-specific config options for it. That would consist of the stuff in the android-base.config, things from the recommended config that you want, and the things from the conditional config (XML) that make sense for your platform. This Android-specific config fragment would need to be maintained moving forward, so someone needs to keep an eye on the kernel/configs project for changes. You'll find out via a warning at boot if you're missing a required config, but better to be informed ahead of time by watching the kernel/configs repo.

Patch-set: 4
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>