add build.config's for upstream LTS branches
As we cannot store build.config files (used for build/build.sh) within
the upstream repositories (they are plain upstream mirrors), provide a
configuration hierarchy here and refer to them from the kernel/manifest
for the branch.

The build.config files can be used to build upstream kernel branches
with build/build.sh after acquiring sources (and toolchain!) with repo.

The background behind this change is to ensure stable toolchains and
configurations when building kernels. That is useful for ABI-checked
builds.

Change-Id: I44a82140713614bfa870ecf1b48c8a1f87305a48
Signed-off-by: Matthias Maennich <maennich@google.com>