Update patch set 2
Patch Set 2:

Not sure if this is part of your plan already, but since you are relying on "=m", you'll also need a VTS test which ensures the module is present on the device. Just checking ikconfig is not a sufficiently strong guarantee that the device manufacturer installed the module.

Patch-set: 2
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>