Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> What does it mean for a device not to have RTL8152 support?
> This is a usb dongle with the driver written in C in the upstream linux kernel.

It mean that android device could work without it. Why should we compile it if it is not to be used? There is no word about RTL8152 in android CDD . You could support it, but this driver should not be a mandatory one.

Patch-set: 2
Reviewer: Gerrit User 1447368 <1447368@85c56323-6fa9-3386-8a01-6480fb634889>