Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> What does it mean for a device not to have RTL8152 support?
> This is a usb dongle with the driver written in C in the upstream linux kernel.

Let me explain our case more precisely. We are porting Android to the automotive board that contains dedicated Ethernet port, so chances that users will use USB-to-Ethernet converters are pretty low. Also we are trying to minimize the kernel size and compile only those drivers that needed for given platform. The motivation to include this specific driver to the kernel is not clear to us. If supporting of USB-to-Ethernet converters is mandatory for Android, then we may need to add also other drivers. If not, then this driver has to be removed.

Patch-set: 2
CC: Gerrit User 1559352 <1559352@85c56323-6fa9-3386-8a01-6480fb634889>