Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I bought nearly a dozen different gigabit usb dongles from various manufacturers and they all used this driver.
> 
> As such this is definitely useful on any Android phone or tablet and probably plenty other Android devices (smart speakers, Android TV, etc...).
> 
> What harm is there in including this as a module?
> It's just a small amount of disk space.

Yes, we can compile it as module for the sake of passing the test. It's not a big issue to add just a 1 module but to reduce the images size and boot time, we are following the policy of adding only those modules that required by CDD. Since we couldn't find such requirement in CDD that caused the confusion. 

I agree that USB to Ethernet could be useful especially for those devices that have no Ethernet ports, but this is not the case for the automotive boards that usually have the dedicated Ethernet port. That is why I think that this requirement should be optional.

Patch-set: 2