Update patch set 3
Patch Set 3:

What is the problem with building this driver as a module?

Are you out of space that this driver being removed give you back?

If you have USB (and automotive does), don't you need/want this anyway?

I don't think we have a way to do "only for automotive this is not needed", right?  So by removing this now, you are taking away the test for all handheld devices where we do want this option to be present, which isn't ok.

Patch-set: 3
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>