Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What is the problem with building this driver as a module?
> 
> Are you out of space that this driver being removed give you back?
> 
> If you have USB (and automotive does), don't you need/want this anyway?
> 
> I don't think we have a way to do "only for automotive this is not needed", right?  So by removing this now, you are taking away the test for all handheld devices where we do want this option to be present, which isn't ok.

I don't think we should be requiring all vendors to support this use case. We certainly shouldn't prevent it, and we can recommend it by putting it in android-recommended.config. But the required android configs should be for things that are *required* to pass CTS and VTS. Not for optional things.

For example, the recommended configs currently include various USB input devices (keyboards, joysticks, mice), other filesystems, and even cpufreq.

Patch-set: 3
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>