Update patch set 1
Patch Set 1: Autosubmit+1 Presubmit-Ready+1

It needs to be for 4.19+ only as we will only be CPing in https://android-review.googlesource.com/c/kernel/common/+/1166025 (which represents the upstream commit referenced in the commit message) to 4.19 and mainline (once accepted there, it will be cherry-picked back to 4.19).

This will, IMHO, need to get in _before_ aosp/1166025 gets merged before 11/30. There is no harm for it existence here.

The issue at that point will be _how_ do we enforce this for android-mainline, the growing assumption is once android-mainline become android-5.4 (WAG), that all the configs here will be copied over to the new supported kernel requirements here.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=+1