Update patch set 1
Patch Set 1:

I'm checking with Yifan to see if there's a way to allow IKHEADERS to be either a =y or =m when the vintf check runs. In the interim, set to =y. Sounds good?

I'll update the cuttlefish kernel in a topic with this change once there's been a 5.4 build with the "IKHEADERS to =y" change built into it.

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>