Require that CRYPTO_MD4 not be set
The MD4 digest algorithm is heavily broken and obsolete, and there
shouldn't be any reason for it to be enabled in the kernel's crypto API.
So let's require that it not be enabled.

(Note: explicitly handling MD4 does *not* mean that it's fine to use
every other crypto algorithm.  MD4 is just a particularly obvious case,
which previously was unnecessarily enabled in GKI.  So there was a good
opportunity to make this change together with disabling it in GKI.)

Bug: 150302692
Test: Treehugger
Change-Id: I8884c24d50cf3e33eb5e3b1db46578cd6c3bfaef