Disable CONFIG_IO_URING
This feature is not used by Android. It's new, complex, under
active development, and thus somewhat high risk from a security
point-of-view [1][2]. It also has no access controls that we can use
to limit which processes may use it. Let's disable it until we have
actual use cases and a mechanism such as LSM-hooks for limiting
access[3].

[1] https://www.graplsecurity.com/post/iou-ring-exploiting-the-linux-kernel
[2] https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2021-41073
[3] Only a select few processes on Android use seccomp. Enabling it
    comes at a performance cost. In cases like this, it can be better
    to disable unused kernel functionality. That reduces the kernel's
    size and reduces security risk.

Bug: 201846396
Test: TH
Change-Id: I748ba9ec7e1d6beae8c14cfee7ce08b8ba15ad7b