p/android-4.14 - require CONFIG_NETFILTER_XT_MATCH_OWNER=y instead of QTAGUID
Technically this is not a valid change, as it restricts valid
configurations of the kernel, but... in practice... this is required
already by at least Android 11 (R) in order for traffic accounting
to work correctly.

When using xt_qtaguid Android R CTS fails:
  testTrafficStatsForLocalhost fails "java.lang.IllegalArgumentException: lower must be less than or equal to upper"
due to an extremely subtle difference in the implementation of
'-m owner --uid-owner XXX' (which xt_qtaguid breaks, by
effectively making the condition always true).

This was reported by Nvidia back in August 2021 in bug 196522401.

It takes a lot of painful debugging before we get to comment 36
which tries to make sense of things... and then comment 38 gets
to the root of the problem.

Bug: 196522401
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I883e0c521e615a99f08361328dde2cd0d9a3d861