Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4873ab7e-2108-4a6b-9f36-6544e540137b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I410aec1f59621bc5af02b0a1dadd8f87c77a135c