Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b5c7eb3-eeab-49c9-838b-f48b6288d66a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4bdf0f6dd89b769ec3e63653a60d272b5c4158b