Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b54dbdd-88e9-494d-aaf8-1cd586ae8317). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bebc06d25c50039972fa5d66f0dd06f85e7b396