Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40bbdbc7-441e-4fb7-8f7e-5555a7c503d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52bbaa00e878e759a5b5151a24b0c248eab88e29