Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f4d4470-02d0-4525-bbd7-435cb6d17f29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f9d9e55ad33f92c7744c8320f936eab5717f58a