Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fdc88db-15dc-470f-a3b7-93048a77f859). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If87595ce5a4342d6f5a4fed46cf888c6dccc3b52