Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4ff7457-6389-4f44-b6d2-f4dd5fff16ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I66ba954d3a345f7727b4421e605f19a327958a20