Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82e5cf81-6085-4762-abf7-2abad741ba83). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib06cc5fb7382a1b8c9ae5162697262a721802a9a