Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d0a1cbb-604f-431d-8cb8-b64dd3e44e1f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia3fbabc73b4aad43ee4817b3d50845746cb126a7