Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d27bf36-7899-4fac-a6d8-5314eaccbf00). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib52ce8187a6b648126c945b926cee41afa075e13