Revert "Add android-6.1 for Android 14"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f3fafd6-f756-4b17-9bd4-0af0eec27889). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib02f6b90b24e70767532cf5c3dea07f29080e469