Revert "Move current requirements to v/ folder."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00c49956-c1c7-45fd-83a2-3a3d807c7399). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e3890e1eb779323854eefd18c99976f178b3e2f