Revert "Move current requirements to v/ folder."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2267824a-69bf-46fc-bfe1-3758030457aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4897cab2b54f297d24a7016dfc7aba7bcc7be380