Update freeze.py for new scheme.
We no longer keep a current version of the configs. We only
have the dessert letter for the next release.

Update the script to reflect this change.

In addition, the script no longer updates hardware/interfaces;
it is up to that repository to update necessary definitions.

In addition, per guideline for scripts for advancing to
the next release, delete all git commands in the script.

Test: manual (freeze.py v w)
Bug: 279809082

Change-Id: Ib6f6fa087e4830abdfe97f2f8c395a887a418021