Revert "Remove Android Q kernel configs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/264ce28c-155d-49da-b40b-7ccdf2bef836). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2abdf52d79a12fa1447b4276f77dd1d6c7984cc2