configs: fix insecure config of eBPF
The baseconfig for kernel enables eBPF by setting
“CONFIG_BPF_SYSCALL=y”， leaving “CONFIG_BPF_JIT_ALWAYS_ON”
and “CONFIG_BPF_UNPRIV_DEFAULT_OFF” not set. If the kernel is
compiled with this defconfig, a normal user without any
capabilities at all can load eBPF programs (SOCKET_FILTER type),
and uses the interpreter. Due to the threat of side-channel
attacks and inextirpable mistakes in the verifier, this is
considered insecure, so the patch set these 2 entries as 'y'.

On the other hand, we found that such configs facilitate kernel
bug exploitation. Specifically, an attacker can leverage existing
CVEs to corrupt eBPF prog-array map, hijacking a bpf_prog pointer
(ptrs[xx]) to point to a forged BPF program. In this way, arbitrary
bytecode execution can be achieved, we have proved this concept
with various CVEs(e.g. CVE-2018-18445). Such an attack enhances the
exploitability of CVEs, and is more dangerous than side-channel
threats.

Change-Id: I93f7260ab574ae694649865eef5ef4b9bc50c4b6
Signed-off-by: liboti <liboti3939@gmail.com>